requirements.txt
Pillow~=10.1.0
keyring~=24.2.0
requests~=2.31.0

pyzbar~=0.1.9
opencv-python~=4.8.1.78
numpy~=1.26.1